Feat: Add CharacterSetUp and nextBtn
