+++ src/pages/Settings/CharacterSetUp.module.scss
